fix: 중복 네이밍 수정 및 누락 변경사항 반영
- Domain Article 과 UI 모듈 Article 네이밍 충돌
   -> UI 모듈 Article은 임시 타입으로 변형 또는 제거 예정
- id 타입은 UUID에서 Int로 변경됨 (누락사항)
