+++ Targets/UserInterface/Sources/ContentView.swift
@@ -6,6 +6,8 @@
 //  Copyright © 2022 nyongnyong. All rights reserved.
 //
 
+import DomainInterface
+
 import SwiftUI
 
 public struct ContentView: View {
